# src/main/java/de/project/cinemaapi/generated/api/PutMovieApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.33).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package de.project.cinemaapi.generated.api;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.Schema;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-03-10T12:00:26.623Z[GMT]")
@Validated
public interface PutMovieApi {

    @Operation(summary = "Puts a movie", description = "Saves a new movie", tags={ "admin" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "New movie has been saved"),
        
        @ApiResponse(responseCode = "404", description = "New movie could not be saved") })
    @RequestMapping(value = "/putMovie/{moviename}",
        method = RequestMethod.PUT)
    ResponseEntity<Void> putMovie(@Parameter(in = ParameterIn.PATH, description = "Name of the new movie", required=true, schema=@Schema()) @PathVariable("moviename") String moviename);

}